% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_llosa.R
\name{gen_llosa}
\alias{gen_llosa}
\title{gen_llosa}
\usage{
gen_llosa(
  dataset,
  borne = FALSE,
  annotate = TRUE,
  annotatetext = c("Secondary- \\n \\"Little Influence\\"",
    " Plus - \\n \\"Satisfaction\\"", "Basic - \\n \\"Disatisfaction\\"",
    "Key - \\n \\"Regardless\\"")
)
}
\arguments{
\item{dataset}{the dataset to use}

\item{borne}{booleen do you want xlim and ylim}

\item{annotate}{booleen do you want annotation}

\item{annotatetext}{annotation text}
}
\description{
plot the Llosa matrix
}
\examples{


library(tetraclasse)
gen_satis(100) \%>\%
  prepare_base() \%>\%
  gen_llosa()


}
