% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_llosa.R
\name{gen_satis}
\alias{gen_satis}
\title{gen_satis}
\usage{
gen_satis(N = 500)
}
\arguments{
\item{N}{number of person}
}
\description{
generate some random data for a satisfaction survey
}
\examples{
base <- gen_satis(50)
}
