% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_package.R
\docType{package}
\encoding{UTF-8}
\name{tetraclasse}
\alias{tetraclasse}
\title{tetraclasse}
\description{
Satisfaction Analysis - Tetraclass Model and Llosa Matrix.
}
\examples{


library(tetraclasse)


# generate an example dataset with 150 respondents
base <- gen_satis(150)

#prepare the dataset
base <- prepare_base(base)

# construct the Llosa matrix
gen_llosa(base)
 if(require("dplyr")){
gen_satis(5000) \%>\%
  sample_n(50,replace = TRUE)\%>\%
  prepare_base() \%>\%
  gen_llosa()



gen_satis(500) \%>\%
  prepare_base() \%>\%
  gen_llosa()

}



}
