% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_llosa.R
\name{prepare_base}
\alias{prepare_base}
\title{prepare_base}
\usage{
prepare_base(base, satis_col = "satis")
}
\arguments{
\item{base}{the base to reshape}

\item{satis_col}{satisfaction column name}
}
\description{
reshape the base
}
\examples{
base <- gen_avis(150)
prepare_base(base)
}
