% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_package.R
\docType{package}
\encoding{UTF-8}
\name{tetraclasse}
\alias{tetraclasse}
\alias{tetraclasse-package}
\title{tetraclasse}
\description{
Satisfaction Analysis - Tetraclasse Model And Llosa Matrix.
}
\examples{


library(tetraclasse)

 if(require("dplyr")){
gen_avis(5000) \%>\%
  sample_n(50,replace = TRUE)\%>\%
  prepare_base() \%>\%
  gen_llosa()



gen_avis(500) \%>\%
  prepare_base() \%>\%
  gen_llosa()

}



}

